+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.6.2] - Unreleased
+### Added
+- Button to logout.
+- Allow to upload files in subfolders.
+
+### Changed
+- Improved footer styles.
+
 ## [0.6.1] - 2024-09-12
 ### Changed
 - UI and behavior of the `current-datetime` field [#23].
@@ -372,6 +380,7 @@ First version
 [#22]: https://github.com/lumeland/cms/issues/22
 [#23]: https://github.com/lumeland/cms/issues/23
 
+[0.6.2]: https://github.com/lumeland/cms/compare/v0.6.1...HEAD
 [0.6.1]: https://github.com/lumeland/cms/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/lumeland/cms/compare/v0.5.10...v0.6.0
 [0.5.10]: https://github.com/lumeland/cms/compare/v0.5.9...v0.5.10